Updated to version 1.1
New Features:

Register page
Linked via login page - a new user can now register and is automatically
logged in

Not Found page - used routing to send user to this page in the event of
a route which doesn't exist in the application

Settings Menu (Navigation Bar)
Implemented a settings menu in the navigation bar, currently only the
last option (logout) works.
Used Dropdown.jsx common component

Task - Status
Added dropdown (using Dropdown.jsx) to display and choose each task's
current status.

Task - Delete
Added ability to delete a task

Tasks - Sorting
Added state and sorting logic (using lodash npm) to allow user to sort
displayed tasks according to title / status / updatedAt properties in
ascending or descending order
Created SortTasks.jsx component which uses Dropdown.jsx
to display sort options (properties)

New Common Components:

Dropdown, DropdownMenu, DropdownItem
These components work together to create a dropdown button and a menu
with any number of items.
The Dropdown component handles logic for opening and closing as well as
making use of the custom useClickOutside hook which, when clicking
outside of the menu, closes it.

Form
A new common component which takes an array inputs
([{name, type, placeholder}]) to render a form.
This component also handles form logic in one place such as error
handling, change handling, validation.
Takes children to be rendered bellow submit button.

ProtectedRoute
This component takes a component as a child, it should be used with the
Route.jsx component.
If a user is not logged in (determined via user context) the child will
not render and instead the user is re-routed to the login page.

Custom Hook:

useClickOutside()
This hook takes a handler function as an argument.
It uses the useRef() hook to create a DOM node which is returned and is
meant to be used on an html element via the ref attribute.
It then uses the useEffect function to listen to the mousedown event and
determine if it occured outside of the html element, if so the handler
function is called.

Updated Components:

Task
Using TaskContext instead of prop drilling from Tasks.jsx to this
component.
Refactored getEdited() helper functions into /utilities/dateUtils.js

TaskTextarea
Used instead of TextArea common component since it now has special
features specific to tasks.
It renders a div when not a selected task, also highilights text
accrording to the search query.

TasksGrid
Greatly refactored and minimized simply by using TaskContext instead of
prop drilling.

common/Button
When passed 'icon' prop it only renders the icon effectively becoming an
"icon button".
When not passed the 'icon' prop it uses its children / label property as
a labeled button. Also accepts the 'leftIcon' and 'rightIcon' props to
display icons to the sides of the label.

Context:

TasksContext
Passes down data from Tasks.jsx down the component tree, mainly used by
Task.jsx which is two levels below.

UserContext
Provided in index.js to be used by routing components (ProtectedRoute),
as well as login page.
Also used within app by SettingsMenu in NavBar