Continued Refactoring for tasks functionality
This refactoring primarlity causes redux to reflect the state of our
database more consistently.
It also helps by dividing responsiblities of selected and unselected
tasks more clearly into two separate components.

- Redux
Redux now consistently reflects database.
Newly added tasks are no longer added via local redux changes but by
react components (SelectedTask).
Task property changes are also no longer done via redux but rather via
react state.

Added ability to set synced to true / false via action creator.
Since task property updates are now done via state react components can
now control synced.

Added /pending cases for createTask and updateTask async actions for
optimistic updates.

Changed impementation of task update requests.
As per changes to backend which allow for partial data to be sent instead
of an entire task object so too can redux handle this.

Created new getTaskById selector for use with new SelectedTask component
which returns a task according to a given id, if not found returns undefined.

- Tasks component
Changed implementations of adding new tasks.
Added state for storing starter properties for new tasks.
When handling a new task sets these starter properties (which are then
passed down) and also sets the selectedTaskId to an iterating ref called
newTaskId.
The SelectedTask component is rendered on the condition that selectedTaskId
is truthy, if it cannot find a relevant task in the store via a selector
it initializes with an empty state.

Also queries are not cleared upon selecting tasks as the new SelectedTask
component is not affceted by filters any longer.

- TasksSelected component
Now accepts selectedTaskId and newTaskProperties object.
Uses the new SelectedTask component and passes on these props to render it.
SelectedTask uses CSSTransitions, its "in" condition property is the
boolean value of selectedTaskId which determines whether to render it
or not.

- SelectedTask component
New component which returns a Task component but injects it with data it
holds within state.
Essentialy a higher order component which wraps a Task component.

Here we have state called "data" which is initialized depending on the
given taskId prop.
If a task exists in the store with the given taskId we initialize state
with its properties, otherwise we initialize an empty state and also add
stateingProperties which are also given as a prop (intended for new tasks).

Here we also take care of debounce saves whenever either title / content
properties of data are changed.
Debouncing only occurs when a task isn't new.

Here we also handleExit which saves a task if it isn't new but has been
edited or if it is new but isn't empty. We call this funtion when a user
cicks outside use useClickOutside hook.
We pass this function to the Task component to be used in the footer section.

This component holds a handleChange function which takes e.currentTarget
and sets it within the state.

We render the Task component and pass it the selected prop alongside
isNew which determines whether it's a new task, onChange and onExit, and
we pass the data state as the task prop.

- Task
This component now recieves the selected prop strictly from the SelectedTask
component and affect certain functionality as well as styling.
We no longer use TaskContext but pass everything via props.

When a task isn't selected it is fed data via the task prop directy from
the store, and changes made to it while not selected are made directly to
the store via async actions which affect the database (no more debounce).

When a task is selected debouncing capabilities are handled via the SelectedTask
component which wraps it.

A major implemtation change has to do with handleUpdateChange which sends
an update action with only relevant data for properties which are not inputted
as text (labels / status / priority).
When a task isn't selected we can assume it isn't new, this means we send
the update action, redux then feeds back these changes to the task when
rendering it within a grid or table, here we see the benfits of optimistic
updates via redux.
When a task is selected we also perform this request but only if it isn't a new task.
When a task is selected we also call the onChange function passed down
from SelectedTask to reflect changes in the state as any changes to the store
via an update should not affect the state of the SelectedTask component.